+++ App/Controllers/reservation.controller.js
@@ -1,28 +1,27 @@
-import decodeToken from '../Middleware/decodeToken.js'
-import Reservations from '../Models/order.model.js'
-import ReservationLines from '../Models/orderline.model.js';
-import SeatModel from '../Models/seat.model.js';
-import EventModel from '../Models/event.model.js';
-import StageModel from '../Models/stage.model.js';
-import { QueryParamsHandle } from '../Middleware/Helpers.js';
+import Reservations from '../Models/reservation.model.js'
+import ReservationLines from '../Models/reservationline.model.js';
+import Seats from '../Models/seat.model.js';
+import Events from '../Models/event.model.js';
+import Stages from '../Models/stage.model.js';
+import { QueryParamsHandle } from '../../Middleware/helpers.js';
 
 // Sætter modellers relationelle forhold - een til mange
 Reservations.hasMany(ReservationLines)
 ReservationLines.belongsTo(Reservations)
 
 // Sætter modellers relationelle forhold - een til mange
-SeatModel.hasMany(ReservationLines)
-ReservationLines.belongsTo(SeatModel)
+Seats.hasMany(ReservationLines)
+ReservationLines.belongsTo(Seats)
 
 // Sætter modellers relationelle forhold - een til mange
-EventModel.hasMany(Reservations)
-Reservations.belongsTo(EventModel)
+Events.hasMany(Reservations)
+Reservations.belongsTo(Events)
 
 // Sætter modellers relationelle forhold - een til mange
-StageModel.hasMany(EventModel)
-EventModel.belongsTo(StageModel)
+Stages.hasMany(Events)
+Events.belongsTo(Stages)
 
-class OrderController {
+class ReservationController {
 	/**
 	 * List Metode - henter alle records
 	 * @param {object} req 
@@ -32,13 +31,19 @@ class OrderController {
 	 list = async (req, res) => {
 		const qp = QueryParamsHandle(req, 'id, firstname')
 
-		const result = await Reservations.findAll({
-			order: [qp.sort_key],
-			limit: qp.limit,
-			attributes: qp.attributes
-		})
-		// Parser resultat som json
-		res.json(result)
+		try {
+			const result = await Reservations.findAll({
+				order: [qp.sort_key],
+				limit: qp.limit,
+				attributes: qp.attributes
+			})
+			// Parser resultat som json
+			res.json(result)				
+		} catch (error) {
+			res.status(418).send({
+				message: `Something went wrong: ${error}`
+			})						
+		}
 	}
 
 	/**
@@ -47,35 +52,49 @@ class OrderController {
 	 * @param {object} res 
 	 * @return {object} Returnerer JSON object med detaljer
 	 */
-	 get = async (req, res) => {
-		// Sætter resultat efter sq metode
-		const result = await Reservations.findOne({
-			attributes: ['firstname', 'lastname', 'address', 'zipcode', 'city', 
-							'email', 'created_at'
-			],
-			include: [
-				{
-					model: EventModel,
-					attributes: [['id','event_id'], 'title', 'price'],
-					include: {
-						model: StageModel,
-						attributes: ['name']
-					}
-				},
-				{
-					model: ReservationLines,
-					attributes: ['seat_id'],
-					include: {
-						model: SeatModel,
-						attributes: ['row', 'number']
-					}
-				}
-			],
-			// Where clause
-			where: { id: req.params.id}
-		});
-		// Parser resultat som json
-		res.json(result)
+	 details = async (req, res) => {
+		const { id } = req.params
+
+		if(id) {
+			try {
+				// Sætter resultat efter sq metode
+				const result = await Reservations.findOne({
+					attributes: ['firstname', 'lastname', 'address', 'zipcode', 'city', 
+									'email', 'created_at'
+					],
+					include: [
+						{
+							model: Events,
+							attributes: [['id','event_id'], 'title', 'price'],
+							include: {
+								model: Stages,
+								attributes: ['name']
+							}
+						},
+						{
+							model: ReservationLines,
+							attributes: ['seat_id'],
+							include: {
+								model: Seats,
+								attributes: ['number']
+							}
+						}
+					],
+					// Where clause
+					where: { id: id}
+				});
+				// Parser resultat som json
+				res.json(result)
+			} catch (error) {
+				res.status(418).send({
+					message: `Something went wrong: ${error}`
+				})										
+			}
+		} else {
+			res.status(403).send({
+				message: 'Wrong parameter values'
+			})
+		}
 	}
 	/**
 	 * Create Metode - opretter ny record
@@ -84,21 +103,40 @@ class OrderController {
 	 * @return {number} Returnerer nyt id
 	 */
 	 create = async (req, res) => {
-		req.body.user_id = decodeToken(req).user_id
 		const { firstname, lastname, address, zipcode, city, seats } = req.body
 		const lines = []
 
 		if(firstname && lastname && address && zipcode && city) {
-			const model = await Reservations.create(req.body)
-			seats.map(async seat => {
-				const newline = await ReservationLines.create({seat_id: seat, order_id: model.id})
-				lines.push(newline.id);
-				
-			})
-			console.log(lines);
-			return res.json({newId: model.id, lines: lines})
+
+			try {
+				const model = await Reservations.create(req.body)
+
+				await Promise.all(seats.map(async seat => {
+					try {
+						const newline = await ReservationLines.create({
+							seat_id: seat, 
+							reservation_id: model.id
+						})
+						lines.push(newline.dataValues.id);	
+					} catch (err) {
+						console.error('Kunne ikke oprette reservationslinje')
+					}
+				}))
+
+				return res.json({
+					message: `Record created`,
+					newId: model.id,
+					// reservation_line_ids: lines
+				})
+			} catch (error) {
+				res.status(418).send({
+					message: `Could not create record: ${error}`
+				})														
+			}
 		} else {
-			res.sendStatus(418)
+			res.status(403).send({
+				message: 'Wrong parameter values'
+			})
 		}
 	}	
 
@@ -109,15 +147,26 @@ class OrderController {
 	 * @return {boolean} Returnerer true/false
 	 */	
 	 update = async (req, res) => {
+		const { id } = req.params
 		const { firstname, lastname, address, zipcode, city } = req.body
 
 		if(firstname && lastname && address && zipcode && city) {
-			const model = await Reservations.update(req.body, {
-				where: {id: req.params.id}
-			})
-			return res.json({status: true})
+			try {
+				const model = await Reservations.update(req.body, {
+					where: {id: id}
+				})
+				return res.json({
+					message: `Record updated`
+				})
+			} catch (error) {
+				res.status(418).send({
+					message: `Could not update record: ${error}`
+				})																		
+			}
 		} else {
-			res.send(418)
+			res.status(403).send({
+				message: 'Wrong parameter values'
+			})
 		}
 	}
 
@@ -128,20 +177,32 @@ class OrderController {
 	 * @return {boolean} Returnerer true/false
 	 */	
 	remove = async (req, res) => {
-		try {
-			await ReservationLines.destroy({ 
-				where: { order_id: req.params.id }
+		const { id } = req.params
+
+		if(id) {
+			try {
+				await ReservationLines.destroy({ 
+					where: { reservation_id: id }
+				})
+				await Reservations.destroy({ 
+					where: { id: id }
+				})
+				res.status(200).send({
+					message: `Record deleted`
+				})
+			}
+			catch(error) {
+				res.status(418).send({
+					message: `Could not delete record: ${error}`
+				})																		
+			}	
+		} else {
+			res.status(403).send({
+				message: 'Wrong parameter values'
 			})
-			await Reservations.destroy({ 
-				where: { id: req.params.id }
-			})
-			res.sendStatus(200)
-		}
-		catch(err) {
-			res.send(err)
 		}
 	}		
 
 }
 
-export { OrderController }
+export default ReservationController
