+++ Core/Controllers/group.controller.js
@@ -15,14 +15,20 @@ class GroupController {
 		// Indhenter parametre fra request objekt
 		const qp = QueryParamsHandle(req, 'id, name')
 
-		// Eksekverer sequelize metode med management values
-		const result = await Groups.findAll({
-			attributes: qp.attributes,
-			order: [qp.sort_key],
-			limit: qp.limit
-		})
-		// Udskriver resultat i json format
-		res.json(result)
+		try {
+			// Eksekverer sequelize metode med management values
+			const result = await Groups.findAll({
+				attributes: qp.attributes,
+				order: [qp.sort_key],
+				limit: qp.limit
+			})
+			// Udskriver resultat i json format
+			res.json(result)			
+		} catch (error) {
+			res.status(418).send({
+				message: `Could not get group list: ${error}`
+			})												
+		}
 	}
 
 	/**
@@ -33,13 +39,26 @@ class GroupController {
 	details = async (req, res) => {
 		// Destructure assignment af id. 
 		const { id } = req.params || 0
-		// Eksekverer sequelize metode med attributter og where clause
-		const result = await Groups.findOne({
-			attributes: ['id', 'name', 'description', 'is_active', 'createdAt', 'updatedAt'],
-			where: { id: id }
-		})
-		// Udskriver resultat i json format
-		res.json(result)
+
+		if(id) {
+			try {
+				// Eksekverer sequelize metode med attributter og where clause
+				const result = await Groups.findOne({
+					attributes: ['id', 'name', 'description', 'is_active', 'createdAt', 'updatedAt'],
+					where: { id: id }
+				})
+				// Udskriver resultat i json format
+				res.json(result)
+			} catch (error) {
+				res.status(418).send({
+					message: `Could not get group details: ${error}`
+				})					
+			}	
+		} else {
+			res.status(403).send({
+				message: 'Wrong parameter values'
+			})
+		}
 	}
 
 	/**
@@ -52,12 +71,23 @@ class GroupController {
 		const { name, description, is_active } = req.body;
 		// Tjekker felt data
 		if(name && description && is_active) {
-			// Opretter record
-			const model = await Groups.create(req.body)
-			// Sender nyt id som json object
-			res.json({ newId: model.id })
+			try {
+				// Opretter record
+				const model = await Groups.create(req.body)
+				// Sender nyt id som json object
+				res.json({
+					message: `Record created`,
+					newId: model.id
+				})				
+			} catch (error) {
+				res.status(418).send({
+					message: `Could not create record: ${error}`
+				})																						
+			}
 		} else {
-			res.sendStatus(418)
+			res.status(403).send({
+				message: 'Wrong parameter values'
+			})
 		}
 	}
 
@@ -73,16 +103,24 @@ class GroupController {
 		const { name, description, is_active } = req.body;
 		// Tjekker felt data
 		if(id && name && description && is_active) {
-			// Opretter record
-			const model = await Groups.update(req.body, {
-				where: { id: id }
-			})
-			// Sender nyt id som json object
-			res.json({ 
-				msg: 'Record update' 
-			})
+			try {
+				// Opretter record
+				const model = await Groups.update(req.body, {
+					where: { id: id }
+				})
+				// Sender nyt id som json object
+				res.json({ 
+					message: 'Record updated' 
+				})				
+			} catch (error) {
+				res.status(418).send({
+					message: `Could not update record: ${error}`
+				})																										
+			}
 		} else {
-			res.sendStatus(418)
+			res.status(403).send({
+				message: 'Wrong parameter values'
+			})
 		}	
 	}
 
@@ -93,14 +131,25 @@ class GroupController {
 	 */	
 	remove = async (req, res) => {
 		const { id } = req.body
-		try {
-			await Groups.destroy({ 
-				where: { id: id }
+
+		if(id) {
+			try {
+				await Groups.destroy({ 
+					where: { id: id }
+				})
+				res.status(200).send({
+					message: `Record deleted`
+				})
+			}
+			catch(err) {
+				res.status(418).send({
+					message: `Could not delete record: ${error}`
+				})																										
+			}	
+		} else {
+			res.status(403).send({
+				message: 'Wrong parameter values'
 			})
-			res.sendStatus(200)
-		}
-		catch(err) {
-			res.send(err)
 		}
 	}	
 }
